requirement.txt
asgiref==3.2.7
Django==3.0.4
django-extensions==2.2.8
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
graphviz==0.13.2
Markdown==3.2.1
Pillow==7.0.0
psycopg2==2.8.4
pydotplus==2.0.2
PyJWT==1.7.1
pyparsing==2.4.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
